DB-8714 Factor out common SpliceDefaultCompactor functions
With this commit, we only factor out the functions of
SpliceDefaultCompactor which are absolutely identical accross all
hbase_sql/hbase<...> versions.

A later commit may be necessary to factor out the meaty part as well.
